+++ package.json
@@ -1,56 +1,45 @@
 {
-  "name": "Kingsmen732",
-  "version": "0.1.0",
+  "name": "zapfolio",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
+  },
   "dependencies": {
-    "@bentoproject/accordion": "^1.2302271541.0",
-    "@fortawesome/fontawesome-free": "^6.6.0",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "crypto-browserify": "^3.12.0",
-    "dotenv": "^16.4.5",
-    "framer-motion": "^11.1.1",
-    "gsap": "^3.12.5",
-    "lucide-react": "^0.371.0",
+    "@fortawesome/fontawesome-free": "^7.1.0",
+    "@tailwindcss/vite": "^4.1.14",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "crypto-browserify": "^3.12.1",
+    "dotenv": "^17.2.3",
+    "framer-motion": "^12.23.22",
+    "gsap": "^3.13.0",
+    "lucide-react": "^0.545.0",
     "os-browserify": "^0.3.0",
     "path-browserify": "^1.0.1",
-    "react": "^17.0.2",
+    "react": "^19.2.0",
     "react-awesome-slider": "^4.1.0",
-    "react-dom": "^17.0.2",
-    "react-icons": "^5.1.0",
-    "react-scripts": "5.0.1",
-    "react-typical": "^0.1.3",
-    "swiper": "^11.1.14",
-    "web-vitals": "^2.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "react-dom": "^19.2.0",
+    "react-icons": "^5.5.0",
+    "react-type-animation": "^3.2.0",
+    "swiper": "^12.0.2",
+    "tailwindcss": "^4.1.14",
+    "web-vitals": "^5.1.0"
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.19",
-    "postcss": "^8.4.38",
-    "tailwindcss": "^3.4.3"
+    "@eslint/js": "^9.36.0",
+    "@types/react": "^19.1.16",
+    "@types/react-dom": "^19.1.9",
+    "@vitejs/plugin-react": "^5.0.4",
+    "eslint": "^9.36.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.22",
+    "globals": "^16.4.0",
+    "vite": "^7.1.7"
   }
 }
