+++ public/manifest.json
@@ -8,7 +8,7 @@
       "type": "image/x-icon"
     },
     {
-      "src": "logo192.png",
+      "src": "./logo192.png",
       "type": "image/png",
       "sizes": "192x192"
     },
