+++ src/App.jsx
@@ -1,0 +1,18 @@
+import { useState } from 'react'
+import reactLogo from './assets/react.svg'
+import viteLogo from '/vite.svg'
+import './App.css'
+import MainContainer from "./components/MainContainer";
+import Slider from './components/slider';
+
+function App() {
+
+
+  return (
+    <div>
+      <Slider/>
+    </div>
+  )
+}
+
+export default App
