+++ src/components/EducationSection.jsx
