+++ src/components/slider.jsx
