+++ src/index.css
@@ -1,29 +1,59 @@
-@import url('https://fonts.googleapis.com/css2?family=Space+Grotesk:wght@300;700;800&display=swap');
+/* Import Google Font */
+@import url('https://fonts.googleapis.com/css2?family=Space+Grotesk:wght@300;500;700&display=swap');
 
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
+@import "tailwindcss";
 
+/* Reset some default styles */
 body {
   margin: 0;
   font-family: 'Space Grotesk', -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   background-color: #f9fafb; /* Light background */
+  color: #111827; /* default text color (dark gray) */
 }
 
+/* Font Weights */
+.font-light {
+  font-weight: 300;
+}
+
+.font-medium {
+  font-weight: 500;
+}
+
+.font-bold {
+  font-weight: 700;
+}
+
+/* Headings */
 h1, h2, h3 {
-  @apply text-gray-900; /* Tailwind utility for text color */
+  color: #111827; /* Equivalent to Tailwind text-gray-900 */
+  margin: 0; /* Optional: remove default margins */
 }
 
+/* Colors */
+.bg-white {
+  background-color: #ffffff;
+}
+
+.text-white {
+  color: #ffffff;
+}
+
+/* Code styling */
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
 }
 
-/* Add hover effect for links */
+/* Links with hover effect */
+a {
+  color: #1d4ed8; /* default link color */
+  text-decoration: none;
+}
+
 a:hover {
-  @apply text-blue-500 underline; /* Tailwind utility for hover effect */
+  color: #2563eb; /* Tailwind blue-500 equivalent */
+  text-decoration: underline;
 }
